+++ apps/api-worker/src/routes/subscriptions.route.ts
@@ -33,17 +33,19 @@ const createSubSchema = z.object({
   startDate: z.string(),
   endDate: z.string(),
   payment: z.object({
-    amountPaid: z.number().positive(),
+    // Todo dinero en centavos enteros (convención Money, ver AGENTS.md).
+    amountPaid: z.number().int().positive(),
     currencyPaid: z.string(),
     exchangeRateApplied: z.string().nullable().optional(),
     paymentMethod: z.string(),
     paymentMethodDetails: paymentMethodDetailsSchema,
     status: z.enum(['processing', 'validated', 'invalid', 'voided']).optional(),
     paymentDate: z.string().optional(),
-    // Desglose fiscal (Fase 0: schema abierto; el servicio lo calcula por
-    // defecto y solo acepta override con taxOverrideReason en Fase 2).
-    subtotal: z.number().positive().optional(),
-    taxTotal: z.number().min(0).optional(),
+    // Desglose fiscal en centavos enteros (Fase 0: schema abierto; el
+    // servicio lo calcula por defecto y solo acepta override con
+    // taxOverrideReason en Fase 2).
+    subtotal: z.number().int().positive().optional(),
+    taxTotal: z.number().int().min(0).optional(),
     taxDetails: z.array(taxDetailSchema).optional(),
     taxOverrideReason: z.string().optional(),
   }),
